tianchi_submit_demo/requirements.txt
configparser==3.8.1
numpy==1.17.2
pandas==0.25.0
scipy==1.3.1
scikit-learn==0.21.3
torchvision==0.4.0
tqdm==4.35.0
tensorboard_logger==0.1.0

